Move LOAD_CLASS_FACTS to compiled_class.rs
